getWindow and sendAction moved into a util.js
